requirements.txt
Django~=5.0.3
djangorestframework~=3.15.1
channels~=4.0.0
wrapt==1.14.0
langflow~=0.6.11
requests~=2.31.0